Install script updates

+++ install.sh
@@ -1,4 +1,13 @@
-#!/bin/bash
+#!/bin/sh
+
+# Error out if anything fails.
+set -e
+
+# Make sure script is run as root.
+if [ "$(id -u)" != "0" ]; then
+  echo "Must be run as root with sudo! Try: sudo ./install.sh"
+  exit 1
+fi
 
 # Delete omxplayer
 apt -y remove omxplayer
